Speed Transition Fixed and Key Signature Changes Implemented
Custom cues were added to each "Speed" music segment to ensure that any transitions were between the same notes in the segment, as opposed to on the beat (which changes the playing notes order temporarily).

Also, logic was added for key signature changes on a global chord track which will be replaced by the harmonic instrument for the track.